+++ lib/bn_api.dart
@@ -2,6 +2,7 @@ library bnapi;
 
 export 'src/objects.dart';
 export 'src/bn_api.dart' show BnApi;
+export 'src/bn_serialized_api.dart' show BnSerializedApi;
 
 // import 'src/new/websocket.dart';